requirements.txt
# Web scraping
playwright
playwright-stealth

# AI/ML
sentence-transformers
faiss-cpu
litellm
google-genai

# Agent framework
crewai
crewai-tools

# LLM/RAG
langchain
langchain-community

# Notifications
python-telegram-bot

# Scheduling
apscheduler

# Web frameworks
streamlit
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy

# PDF processing
pypdf

# Testing
pytest
pytest-asyncio

# Utilities
python-dotenv
httpx
